requirements.txt
pandas>=2.0.0
openpyxl>=3.1.0
xlrd==1.2.0
python-docx>=0.8.11
numpy>=1.24.0
python-dateutil>=2.8.2

